Bugfix with notification code, changed Spaces to empty views so its compatible with older devices

+++ raceyourself/src/main/java/com/raceyourself/raceyourself/game/GameActivity.java
@@ -333,7 +333,7 @@ public class GameActivity extends BaseFragmentActivity {
 
     @Override
     public void onDestroy() {
-        gameService.removeNotifications();
+        if(gameService != null) gameService.removeNotifications();
         super.onDestroy();
     }
